#1 #7 #15 . Se empieza con la cancelacion de trabajos y se adaptan esteticamente el form implicado.

+++ html/user/emListTaskFormAction.php
@@ -12,26 +12,30 @@ $size = (int) count($_REQUEST);
 db_init();
 $user = get_logged_in_user();
 
-echo "Tamano parametrosss  :(" . $size . ")";
+//echo "Tamano parametrosss  :(" . $size . ")";
 
 page_head(tra("List Of Deleted Files "));
-start_table("align=\"center\"");
+start_table();
+row1("Data Deleting.". '9');
 $contador=1;
 foreach($_REQUEST as $key => $value)
 {
-echo "key : " . $key . " value : " . $value . "<br/>";
-	$pathBorrado="../../sample_results/";
-	$fichero0=$pathBorrado . $key . "_0.gz";
-	$fichero1=$pathBorrado . $key . "_1.gz";
-	$salida=shell_exec('rm -f ' . $fichero0 . " " . $fichero1);
+	if($stcmp($key,"cancel")==0){
+		//lanzar comando de borrado.
+	}else{
+		//echo "key : " . $key . " value : " . $value . "<br/>";
+		$pathBorrado="../../sample_results/";
+		$fichero0=$pathBorrado . $key . "_0.gz";
+		$fichero1=$pathBorrado . $key . "_1.gz";
+		$salida=shell_exec('rm -f ' . $fichero0 . " " . $fichero1);
 
-        echo "<tr><td> Deleted File :" . $fichero0 . "</td>";
-        echo "<tr><td> Deleted File :" . $fichero1 . "</td>";
-	$result=mysql_query("delete from user_workunit where workunit_id = " . $value );
-        echo "<tr><td> Deleted Database References </td>";
+		row2("Deleted File", $fichero0);
+		row2("Deleted File", $fichero1);
+		$result=mysql_query("delete from user_workunit where workunit_id = " . $value );
+	}
 	$contador+=1;	
 } 
-echo "FICHEROS BORRADOS!!!";
+row2("", " Deleted Database References ");
 
 
 end_table();
